healthysmiles-backend/requirements.txt
click==8.0.4
cmake==3.24.1.1
dataclasses==0.8
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
Pillow==8.4.0
pkg_resources==0.0.0
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
